01_Data_pipelines/requirements.txt
pyspark==3.3.1
pytest==7.2.1
python_dateutil==2.8.2
apache-airflow==2.5.1
